+++ apps/client/src/app/lazy-data/model/lazy-island-animal-source.ts
@@ -1,0 +1,3 @@
+export interface LazyIslandAnimalSource {
+  "37604": number[];
+}