+++ apps/client/src/app/pages/island-workshop/planning-optimizer.ts
@@ -7,11 +7,11 @@ type Genome = number[];
 
 export class PlanningOptimizer {
 
-  private static POPULATION_SIZE = 200;
+  private static POPULATION_SIZE = 500;
 
-  private static GENERATIONS = 200;
+  private static GENERATIONS = 301;
 
-  private static MUTATION_FACTOR = 1;
+  private static MUTATION_FACTOR = 5;
 
   private static WEEKLY_RESET_DAY = 2;
 
@@ -58,22 +58,28 @@ export class PlanningOptimizer {
   }
 
   public run(): { score: number, planning: CraftworksObject[] }[] {
+    if (this.remainingHoursBeforeReset < 4) {
+      return [];
+    }
     for (let i = 0; i < PlanningOptimizer.GENERATIONS; i++) {
       this.population = this.newGeneration();
+      // if (i % 100 === 0) {
+      //   this.printDebugScore(i);
+      // }
     }
     return uniqBy(this.population, e => JSON.stringify(e))
       .map(individual => {
         return {
           individual,
-          score: this.fitness(individual)
+          score: this.getRealScore(individual)
         };
       }).sort((a, b) => {
         return b.score - a.score;
       })
       .slice(0, 5)
-      .map(({ individual }) => {
+      .map(({ individual, score }) => {
         return {
-          score: this.getRealScore(individual),
+          score: score,
           planning: individual.map(chromosome => {
             return this.objects[chromosome];
           })
@@ -81,7 +87,7 @@ export class PlanningOptimizer {
       });
   }
 
-  private printDebugScore(): void {
+  private printDebugScore(iteration: number): void {
     const scored = this.population
       .map(individual => {
         return {
@@ -92,8 +98,10 @@ export class PlanningOptimizer {
       .sort((a, b) => {
         return b.score - a.score;
       });
+    console.group(`ITERATION #${iteration}`);
     console.log(`AVG SCORE: ${scored.reduce((acc, e) => acc + e.score, 0) / this.population.length}`);
     console.log(`BEST: ${scored[0].score}`);
+    console.groupEnd();
   }
 
   private newGeneration(): Genome[] {
@@ -186,10 +194,10 @@ export class PlanningOptimizer {
   }
 
   private fitness(genome: Genome): number {
-   return this.getRealScore(genome, 100);
+    return this.getRealScore(genome, 1);
   }
 
-  private getRealScore(genome: Genome, bonusWeight = 1):number{
+  private getRealScore(genome: Genome, bonusWeight = 1): number {
     return genome.reduce((acc, chromosome, index) => {
       if (acc.totalTime > this.remainingHoursBeforeReset) {
         return {