+++ apps/client/src/environments/extracts-hash.ts
@@ -1,1 +1,1 @@
-export const extractsHash = `00438556f782fad44e1e0f55b4390cd96fffb909`;
+export const extractsHash = `85e97a1c957c839a3b936509bcb5c6dac80b9e9a`;