+++ apps/client/src/environments/patch-notes.ts
@@ -1,20 +1,12 @@
 export const patchNotes = `### Bug Fixes
 
-* **alarms:** reuse generic alarm button in more places.
-* **db:** fix create alarm button icon on node pages.
-* **db:** fixed trades not loading on NPC pages.
-* **desktop:** remove currencies from Island Sanctuary inventory tracking.
-* **island-sanctuary:** better gathering location for Island Vine.
-* **list:** fix item names not showing in list panel.
-* **list:** fixed requirements popup showing wrong numbers.
-* **lists:** fixed panel mark as done not checking as done in log.
-* **log-tracker:** fixed fish eyes not showing in fishing log tracker.
-* **search:** removed island buildings from query-less search.
-* **sumbarine:** fix max submarine rank.
-* **universalis:** better query for item entries count.
-* **voyages:** add missing voyages from 6.2 update.
+* **collectables:** fix alarm buttons not showing on timed items.
+* **db:** fix Islekeep's Stone Hatchet recipe.
+* **island-workshop:** display error message when no plannings are found.
+* **log-tracker:** fixed selected recipes not resetting on page change.
 
 
 ### Features
 
-* **island:** new in-app workshop table with daily optimizer.`;
+* **allagan-reports:** support for island pasture animal source.
+* **lists:** add support for landmark recipes.`;